Move BlueZGattCharacteristic to dbus
