+++ bless/backends/bluezdbus/service.py
@@ -8,7 +8,7 @@ from txdbus import client  # type: ignore
 from txdbus.objects import DBusObject, DBusProperty  # type: ignore
 from txdbus.interface import DBusInterface, Property  # type: ignore
 
-from .characteristic import BlueZGattCharacteristic  # type: ignore
+from .dbus.characteristic import BlueZGattCharacteristic  # type: ignore
 
 if TYPE_CHECKING:
     from bless.backends.bluezdbus.application import (  # type: ignore
